+++ packages/components/Avatar/src/Avatar.types.ts
@@ -4,7 +4,7 @@ import { IBackgroundColorTokens, IForegroundColorTokens, IBorderTokens, FontToke
 import { BadgeProps, PresenceBadgeProps, BadgeSize } from '@fluentui-react-native/badge';
 import { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
 
-export const JSAvatarName = 'Avatar';
+export const AvatarName = 'Avatar';
 export const AvatarSizesForTokens = [
   'size20',
   'size24',
@@ -112,7 +112,7 @@ export interface AvatarConfigurableProps {
   size?: AvatarSize;
 }
 
-export interface JSAvatarProps extends IViewProps, AvatarConfigurableProps {
+export interface AvatarProps extends IViewProps, AvatarConfigurableProps {
   /**
    * The appearance when `active="active"`
    *
@@ -169,7 +169,7 @@ export interface JSAvatarProps extends IViewProps, AvatarConfigurableProps {
   imageUrl?: string;
 }
 
-export interface JSAvatarTokens extends IBackgroundColorTokens, IForegroundColorTokens, AvatarConfigurableProps, IBorderTokens, FontTokens {
+export interface AvatarTokens extends IBackgroundColorTokens, IForegroundColorTokens, AvatarConfigurableProps, IBorderTokens, FontTokens {
   /**
    * Avatar opacity which is changed depending on `active` prop.
    */
@@ -193,68 +193,68 @@ export interface JSAvatarTokens extends IBackgroundColorTokens, IForegroundColor
   /**
    * Avatar shapes:
    */
-  circular?: JSAvatarTokens;
-  square?: JSAvatarTokens;
+  circular?: AvatarTokens;
+  square?: AvatarTokens;
 
   /**
    * Token for inactive value of `active` prop
    */
-  inactive?: JSAvatarTokens;
+  inactive?: AvatarTokens;
 
   /**
    * Avatar sizes:
    */
-  size20?: JSAvatarTokens;
-  size24?: JSAvatarTokens;
-  size28?: JSAvatarTokens;
-  size32?: JSAvatarTokens;
-  size36?: JSAvatarTokens;
-  size40?: JSAvatarTokens;
-  size48?: JSAvatarTokens;
-  size56?: JSAvatarTokens;
-  size64?: JSAvatarTokens;
-  size72?: JSAvatarTokens;
-  size96?: JSAvatarTokens;
-  size120?: JSAvatarTokens;
+  size20?: AvatarTokens;
+  size24?: AvatarTokens;
+  size28?: AvatarTokens;
+  size32?: AvatarTokens;
+  size36?: AvatarTokens;
+  size40?: AvatarTokens;
+  size48?: AvatarTokens;
+  size56?: AvatarTokens;
+  size64?: AvatarTokens;
+  size72?: AvatarTokens;
+  size96?: AvatarTokens;
+  size120?: AvatarTokens;
 
   /**
    * Avatar colors:
    */
-  neutral?: JSAvatarTokens;
-  brand?: JSAvatarTokens;
-  darkRed?: JSAvatarTokens;
-  cranberry?: JSAvatarTokens;
-  red?: JSAvatarTokens;
-  pumpkin?: JSAvatarTokens;
-  peach?: JSAvatarTokens;
-  marigold?: JSAvatarTokens;
-  gold?: JSAvatarTokens;
-  brass?: JSAvatarTokens;
-  brown?: JSAvatarTokens;
-  forest?: JSAvatarTokens;
-  seafoam?: JSAvatarTokens;
-  darkGreen?: JSAvatarTokens;
-  lightTeal?: JSAvatarTokens;
-  teal?: JSAvatarTokens;
-  steel?: JSAvatarTokens;
-  blue?: JSAvatarTokens;
-  royalBlue?: JSAvatarTokens;
-  cornflower?: JSAvatarTokens;
-  navy?: JSAvatarTokens;
-  lavender?: JSAvatarTokens;
-  purple?: JSAvatarTokens;
-  grape?: JSAvatarTokens;
-  lilac?: JSAvatarTokens;
-  pink?: JSAvatarTokens;
-  magenta?: JSAvatarTokens;
-  plum?: JSAvatarTokens;
-  beige?: JSAvatarTokens;
-  mink?: JSAvatarTokens;
-  platinum?: JSAvatarTokens;
-  anchor?: JSAvatarTokens;
-  burgundy?: JSAvatarTokens;
-  hotPink?: JSAvatarTokens;
-  orchid?: JSAvatarTokens;
+  neutral?: AvatarTokens;
+  brand?: AvatarTokens;
+  darkRed?: AvatarTokens;
+  cranberry?: AvatarTokens;
+  red?: AvatarTokens;
+  pumpkin?: AvatarTokens;
+  peach?: AvatarTokens;
+  marigold?: AvatarTokens;
+  gold?: AvatarTokens;
+  brass?: AvatarTokens;
+  brown?: AvatarTokens;
+  forest?: AvatarTokens;
+  seafoam?: AvatarTokens;
+  darkGreen?: AvatarTokens;
+  lightTeal?: AvatarTokens;
+  teal?: AvatarTokens;
+  steel?: AvatarTokens;
+  blue?: AvatarTokens;
+  royalBlue?: AvatarTokens;
+  cornflower?: AvatarTokens;
+  navy?: AvatarTokens;
+  lavender?: AvatarTokens;
+  purple?: AvatarTokens;
+  grape?: AvatarTokens;
+  lilac?: AvatarTokens;
+  pink?: AvatarTokens;
+  magenta?: AvatarTokens;
+  plum?: AvatarTokens;
+  beige?: AvatarTokens;
+  mink?: AvatarTokens;
+  platinum?: AvatarTokens;
+  anchor?: AvatarTokens;
+  burgundy?: AvatarTokens;
+  hotPink?: AvatarTokens;
+  orchid?: AvatarTokens;
 }
 
 export interface AvatarSlotProps {
@@ -267,20 +267,20 @@ export interface AvatarSlotProps {
   badge: BadgeProps;
 }
 
-export interface JSAvatarState {
+export interface AvatarState {
   showRing: boolean;
   transparentRing: boolean;
   showBadge: boolean;
 }
 
 export interface AvatarInfo {
-  props: JSAvatarProps;
-  state: JSAvatarState;
+  props: AvatarProps;
+  state: AvatarState;
 }
 
-export interface JSAvatarType {
-  props: JSAvatarProps;
+export interface AvatarType {
+  props: AvatarProps;
   slotProps: AvatarSlotProps;
-  tokens: JSAvatarTokens;
-  state: JSAvatarState;
+  tokens: AvatarTokens;
+  state: AvatarState;
 }